refactor: Move UI elements into /components folder
