+++ src/features/EmailMarketing/shared/InternalTools/components/ExamplePrompts/components/ExamplePromptWithNavigation/ExamplePrompt.tsx
@@ -1,5 +1,5 @@
-import Icon from '../../../../../../../shared/ui/Icon/Icon';
-import {IconType} from '../../../../../../../shared/ui/Icon/iconLibrary';
+import Icon from '../../../../../../../shared/ui/components/Icon/Icon';
+import {IconType} from '../../../../../../../shared/ui/components/Icon/iconLibrary';
 import {EmailMarketingActionType} from '../../../../../../store/emailMarketingStoreTypes';
 import {useEmailMarketingState} from '../../../../../../store/useEmailMarketingState';
 import './ExamplePrompt.scss';
