+++ src/features/EmailMarketing/shared/TargetingBuilderHeader/components/GenerateWithAIForm/GenerateWithAIForm.tsx
@@ -3,10 +3,10 @@ import {SubmitAIPrompt} from '../../../../../shared/ai/ChatGPTService'; // Adjus
 import {useEmailMarketingState} from '../../../../store/useEmailMarketingState';
 
 import './GenerateWithAIForm.scss';
-import Button from '../../../../../shared/ui/Button/Button';
-import Icon from '../../../../../shared/ui/Icon/Icon';
+import Button from '../../../../../shared/ui/components/Button/Button';
+import Icon from '../../../../../shared/ui/components/Icon/Icon';
 import {EmailMarketingActionType} from '../../../../store/emailMarketingStoreTypes';
-import {IconType} from '../../../../../shared/ui/Icon/iconLibrary';
+import {IconType} from '../../../../../shared/ui/components/Icon/iconLibrary';
 import {v4 as uuid} from 'uuid';
 import {AIResponse} from '../../../emailMarketingTypes';
