+++ src/features/shared/ui/components/ComboBox/ComboBox.scss
