+++ src/features/shared/ui/components/CurrencyInput/CurrencyInput.scss
