+++ src/features/shared/ui/components/CurrencyInput/CurrencyInput.test.tsx
