+++ src/features/shared/ui/components/DateInput/DateInput.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {formatDateForDisplay} from '../../../EmailMarketing/shared/formatUtils';
+import {formatDateForDisplay} from '../../../../EmailMarketing/shared/formatUtils';
 
 interface DatePickerProps {
   value?: string;
