+++ src/features/shared/ui/components/Icon/Icon.scss
