+++ src/features/shared/ui/components/Icon/Icon.test.tsx
