+++ src/features/shared/ui/components/NumberInputWithLabel/NumberInputWithLabel.scss
