+++ src/features/shared/ui/components/Sticker/Sticker.test.tsx
