+++ src/features/shared/ui/components/Sticker/Sticker.tsx
