+++ src/features/shared/ui/components/VoteUpOrDown/VoteUpOrDown.scss
