+++ src/features/shared/ui/components/Warning/Warning.scss
